Move SchemaObject into submodule
